fix smartRC driver

+++ tasmota/tasmota_xdrv_driver/xdrv_97_smartRC_RCS.ino
@@ -168,7 +168,7 @@ void CmndRfSend(void)
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv97(uint8_t function)
+bool Xdrv97(uint32_t function)
 {
   bool result = false;
